+++ sdk/core/System.ClientModel/src/Message/PipelineMessage.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 using System.ClientModel.Internal;
+using System.IO;
 using System.Threading;
 
 namespace System.ClientModel.Primitives;
@@ -126,6 +127,87 @@ public class PipelineMessage : IDisposable
         }
     }
 
+    /// <summary>
+    /// Returns the response content stream and releases it ownership to the caller.
+    ///
+    /// After calling this method, any attempt to use the
+    /// <see cref="PipelineResponse.ContentStream"/> or <see cref="PipelineResponse.Content"/>
+    /// properties on <see cref="Response"/> will result in an exception being thrown.
+    /// </summary>
+    /// <returns>The content stream, or <code>null</code> if <see cref="Response"/>
+    /// did not have content set.</returns>
+    public Stream? ExtractResponseContent()
+    {
+        if (Response is null)
+        {
+            return null;
+        }
+
+        SetProperty(typeof(PipelineResponse), true);
+
+        switch (Response.ContentStream)
+        {
+            case ResponseShouldNotBeUsedStream responseContent:
+                return responseContent.Original;
+            case Stream stream:
+                Response.ContentStream = new ResponseShouldNotBeUsedStream(Response.ContentStream);
+                return stream;
+            default:
+                return null;
+        }
+    }
+
+    internal class ResponseShouldNotBeUsedStream : Stream
+    {
+        public Stream Original { get; }
+
+        public ResponseShouldNotBeUsedStream(Stream original)
+        {
+            Original = original;
+        }
+
+        private static Exception CreateException()
+        {
+            return new InvalidOperationException("The operation has called ExtractResponseContent and will provide the stream as part of its response type.");
+        }
+
+        public override void Flush()
+        {
+            throw CreateException();
+        }
+
+        public override int Read(byte[] buffer, int offset, int count)
+        {
+            throw CreateException();
+        }
+
+        public override long Seek(long offset, SeekOrigin origin)
+        {
+            throw CreateException();
+        }
+
+        public override void SetLength(long value)
+        {
+            throw CreateException();
+        }
+
+        public override void Write(byte[] buffer, int offset, int count)
+        {
+            throw CreateException();
+        }
+
+        public override bool CanRead => throw CreateException();
+        public override bool CanSeek => throw CreateException();
+        public override bool CanWrite => throw CreateException();
+        public override long Length => throw CreateException();
+
+        public override long Position
+        {
+            get => throw CreateException();
+            set => throw CreateException();
+        }
+    }
+
     #region IDisposable
 
     protected virtual void Dispose(bool disposing)
@@ -135,14 +217,21 @@ public class PipelineMessage : IDisposable
             var request = Request;
             request?.Dispose();
 
-            _propertyBag.Dispose();
-
             var response = Response;
             if (response != null)
             {
+                if (response.ContentStream is not null && !response.IsBuffered &&
+                    TryGetProperty(typeof(PipelineResponse), out object? extracted) &&
+                    extracted is bool contentExtracted)
+                {
+                    response.ContentExtracted = contentExtracted;
+                }
+
                 response.Dispose();
                 Response = null;
             }
+
+            _propertyBag.Dispose();
 
             _disposed = true;
         }
