+++ sdk/core/System.ClientModel/src/Message/PipelineResponse.cs
@@ -39,6 +39,8 @@ public abstract class PipelineResponse : IDisposable
     private byte[]? _contentBytes;
     internal bool IsBuffered => _contentBytes != null;
 
+    internal bool ContentExtracted { get; set; }
+
     public virtual BinaryData Content
     {
         get
@@ -112,7 +114,7 @@ public abstract class PipelineResponse : IDisposable
         ContentStream = bufferStream;
 
         // TODO: Come back and optimize - this is only for POC at this stage.
-        _contentBytes= bufferStream.ToArray();
+        _contentBytes = bufferStream.ToArray();
     }
 
     private static async Task CopyToAsync(Stream source, Stream destination, TimeSpan timeout, CancellationTokenSource cancellationTokenSource)
@@ -126,7 +128,8 @@ public abstract class PipelineResponse : IDisposable
 #pragma warning disable CA1835 // ReadAsync(Memory<>) overload is not available in all targets
                 int bytesRead = await source.ReadAsync(buffer, 0, buffer.Length, cancellationTokenSource.Token).ConfigureAwait(false);
 #pragma warning restore // ReadAsync(Memory<>) overload is not available in all targets
-                if (bytesRead == 0) break;
+                if (bytesRead == 0)
+                    break;
                 await destination.WriteAsync(new ReadOnlyMemory<byte>(buffer, 0, bytesRead), cancellationTokenSource.Token).ConfigureAwait(false);
             }
         }
